% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_simulate.R
\name{dataset-simulation}
\alias{dataset-simulation}
\alias{fd_simulate}
\alias{fd_proportions}
\title{Simulation of datasets.}
\usage{
fd_simulate(params, times, mgen = NULL, breaks = 50, range = NULL,
  noise = NULL, select = c("hists", "Ns"), model = NULL, mean = FALSE)

fd_proportions(params, times, mgen = NULL, model = NULL)
}
\arguments{
\item{params}{For \code{fd_simulate}, A named numeric vector of coefficients
or a matrix with parameter names in columns as would be returned by
\code{coef} applied to an \code{\link[nlme]{nlme}},
\code{\link[nlme]{nlsList}}, \code{\link[stats]{nls}} or
\code{\link{fd_nls}} object or the result of \code{\link{fd_unif}}. For
\code{fd_proportion}, only a named numeric vector is allowed.}

\item{times}{A numeric vector of times, in hours, at which to simulate
sampling.}

\item{mgen}{Maximum number of generations (if \code{NULL}, the default of the
proliferation model is taken).}

\item{breaks}{The breakpoints to use when forming histograms.  If it is a
single number, number of breaks to use.  If it is a numeric vector, it
directly gives the breakpoints in linear fluorescence units.  It is not
possible to specify different breakpoints for different timepoints.}

\item{range}{A two-element numeric vector giving respectively the lower and
  upper breakpoints, in linear fluorescence units, when \code{breaks} is a
  single numeric value.  If not explicitly specified, the range is taken to
  be
\verb{
}\code{exp(mean(m0) + mean(sd0) * 3) * c(1 / 2.0^(mgen + 4), 1)}.}

\item{noise}{A two-element numeric vector that gives the standard deviations
of a white noise to be added respectively to histograms and cell counts,
after FMM transformations apply (see \code{\link{fd_transform}}).}

\item{select}{What \code{Type} the resulting \code{\link{fd_data}} should
feature.}

\item{model}{An \code{\link{fd_model}} object.  In the case of
\code{fd_norm}, mandatory only if \code{mean} was not generated by
\code{\link{fd_unif}}, \code{\link{fd_draw_unif}}, \code{\link{fd_nls}} or
\code{\link{fd_comb}}.}

\item{mean}{If \code{params} is a matrix, whether to average the columns to
give a single set of parameters (\code{TRUE}) or to generate a dataset for
each row (\code{FALSE}), in which case the dataset contains multiple
\code{Individual}, labelled following the row names or \code{indiv_<i>} if
no row name was given.}
}
\value{
\code{fd_simulate} returns an \code{\link{fd_data}} dataset.
\code{fd_proportions} returns a list with the following elements: \itemize{
\item For branching processes: \describe{ \item{\code{live_pop}}{A named list
of matrices.  The element names are the categories, the rows the timepoints
(in the order of argument \code{times}) and the columns the number of
generations (the first column gives generation 0).  The entries give the
proportions of cells in a given number of generation at a given time and the
rows sum to one.} \item{\code{Ns}}{A matrix giving the number of currently
alive cells in each category (columns) for each timepoint (rows), relative to
the \code{initial} argument passed along to \verb{
}\code{\link{fd_proliferation_branching}}.} \item{\code{lost_pop}}{Same as
\code{live_pop} but with the lost population.} \item{\code{Ns_lost}}{Same as
\code{Ns} but with (cumulative) lost cells.} \item{\code{cum_influx}}{A named
list of matrices, following the format of \code{live_pop}. The entries give
the cumulative influxes of cells in a given number of generation at a given
time, relative to the \code{initial} argument passed along to
\code{\link{fd_proliferation_branching}}.} }
\item For Cyton models: same
format but \code{cum_influx} is not available. }
}
\description{
\code{fd_simulate} simulates an \code{\link{fd_data}} dataset.
\code{fd_proportions} gives the proportion of cells in a given number of
generations and other related metrics.
}
\examples{
un <- fd_unif("gaussian", "branching", n=5)
times <- c(2, 4, 6, 12)
sim <- fd_simulate(un, times,
                   select=c("hists", "hists_lost", "Ns", "Ns_lost"))
unique(sim$Individual)
plot(sim, main="Simulation")

fd_proportions(un, times)

un2 <- fd_unif("gaussian", "cyton", n=5)
fd_proportions(un2, times)
}
\seealso{
Other simulation-related entries: \code{\link{fd_predict}},
  \code{\link{random-parameters}}
}
