% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fd_comb}
\alias{fd_comb}
\title{Comb through many constraints using \code{\link{fd_nls}}.}
\usage{
fd_comb(cstr, fmm = "gaussian", proliferation = "branching", data,
  continue = TRUE, mc.cores = 1, file = NULL, globalname = "mdl", ...)
}
\arguments{
\item{cstr}{A list of \code{\link{constraints}} to comb through}

\item{fmm}{A finite mixture model, same format as with
\code{\link{fd_model}}.}

\item{proliferation}{A proliferation model, same format as with
\code{\link{fd_model}}.}

\item{data}{The \code{\link{fd_data}} dataset on which to perform the search.}

\item{continue}{If \code{TRUE} (default), an error in an optimization does
not stop the whole search but simply results in the corresponding entry of
the return value to be \code{NULL}.  If \code{FALSE}, an error is raised
instead.}

\item{mc.cores}{Number of cores to use for parallel search using
\pkg{parallel}. \code{mc.cores > 1} fails on Windows machines.}

\item{file}{Output \code{file} (or file name) for parallel search (during
which the console output is silenced to avoid conflicts).  The output file
can be read in real time on POSIX machines using \code{tail -f <file>}.}

\item{globalname}{Global name to use for storing the \code{\link{fd_model}}
object. See \code{\link{fd_model-functions}} for more details.}

\item{...}{Additional parameters to pass along to \code{\link{fd_nls}}.}
}
\value{
A named list, each entry being either the result of the optimization
  (an \code{fd_nls} object) or \code{NULL} if the optimization failed, using
  the constraints \code{cstr} and in the same order as the constraints
  \code{cstr}.  The names are the names of \code{cstr}, if provided, or
  \code{constr_<i>}, with \code{<i>} the index of the constraint if no name
  was made available.
}
\description{
Comb through many constraints using \code{\link{fd_nls}}.  The result can
then be fed to \code{\link{fd_aictab}}.
}
\section{Required packages}{
 Please install \code{parallel}.
}

\examples{
# Shared constraints
CC <<- FdCommonConstraints
`#macr` <- ~CC$`#noss` + pro:all:{res <- 0} + fmm:{c0 <- 0} + CC$`#mm_xyxz`

control <- list(maxit = 1L)
\dontrun{
control <- list(maxit = 100L, simple.function = TRUE)
}

source(system.file("contrib", "nlsSA.R", package="fluodilution"))
data(FdSTyphimuriumWTC57)
ans <- fd_comb(lapply(list("1100" = CC$`#delta_1100`,
                           "1111" = CC$`#delta_1111`),
                      catcstr, `#macr`),
               data = cutoff(FdSTyphimuriumWTC57),
               control = control,
               nlsfun = nlsSA,
               trace = TRUE)

summary(ans[[1L]])
summary(ans[[2L]])

}
\seealso{
\code{\link{fd_nls}}, \code{\link{fd_aictab}}

Other optimization-related entries: \code{\link{fd_gaussian_fmm_solve}},
  \code{\link{fd_minuslogl}}, \code{\link{fd_nls}}
}
