% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R, R/pretty.R
\name{fd_data}
\alias{fd_data}
\alias{fd_transform}
\alias{cutoff}
\alias{fd_expand}
\alias{plot.fd_data}
\title{Manage a fluorescence dilution dataset.}
\usage{
fd_data(data, categories = NULL, inoculums = NULL, timepoints = NULL,
  na.action = na.pass)

fd_transform(x, hist = "identity", N = "log10")

cutoff(x, cutoff = NULL, threshold = 0.01)

fd_expand(x, length.N = 50, seq = NULL, separate = TRUE, by = NULL)

\method{plot}{fd_data}(x, type = "overview", main = NULL, ...)
}
\arguments{
\item{data}{The original \code{data.frame} (or perhaps a
\code{\link[nlme]{groupedData}}).}

\item{categories}{The specific order for the category levels.  If \code{NULL}
and the \code{Category} field is not yet a factor, the levels will be in
lexicographical order.}

\item{inoculums}{Same with the \code{Inoculum} field.}

\item{timepoints}{Same with the \code{Timepoint} field.}

\item{na.action}{An \code{\link[stats]{na.action}} object to apply to
\code{data}.}

\item{x}{An \code{fd_data} object.}

\item{hist}{A transformation object from the (graphical) package \pkg{scales}
(or a character string converted to a transformation object) and to be
applied to the \code{y} field when \code{Weight == "hist"}.}

\item{N}{A transformation object from the (graphical) package \pkg{scales}
(or a character string converted to a transformation object) and to be
applied to the \code{y} field when \code{Weight == "N"}.}

\item{cutoff}{If specified, overrides the behaviour of the attribute
\code{cutoff} of \code{data}.  If a named vector of multiple entries, it
contains the cutoff points in linear units of fluorescence (the names are
the timepoint identifiers).  If it is a scalar, gives the maximum number of
generations to consider: the cut is then made at \verb{
} \code{exp(m0) /
2^cutoff}, with \code{m0} the log-mean of the sample.}

\item{threshold}{Below this value, proportions in fluorescence histograms are
discarded. Can help with egregious heteroskedasticity problems.}

\item{length.N}{The new number of cell counts per group.}

\item{seq}{Alternatively, the times at which to calculate the cell counts can
be explicitly given.}

\item{separate}{If \code{TRUE} (default), the new timepoints to consider are
found independently for each group.  This is useful to avoid extrapolation
when data is unbalanced.}

\item{by}{The additional columns with which to do the grouping (Grouping is
already made by \code{Category}). If \code{NULL}, \code{separate = TRUE}
and \code{data} inherits \code{groupedData}, this is taken from the
grouping information of \code{data}. Otherwise, no additional grouping is
used.}

\item{type}{A character vector containing the types of plots to show. If
\code{"overview"}, a selection of plots are shown on the same page.
Otherwise, the plots are shown sequentially.  Plots include \code{"hist"}
(fluorescence histograms), \code{"N"} (cell count), \code{"range"} (an
overview of cutoff points and initial log-mean fluorescence and
log-standard deviation), \code{"balancing"} (data balancing across times
and categories), \code{"coverage"} (in terms of number of cells observed
under flow cytometry, if available) and \code{"cutoff"} (the percentage of
the population below the cutoff point).  Additionally, \code{"all"} shows
all these plots at once.}

\item{main}{Plot title when \code{type="overview"}.}

\item{...}{Additional arguments to the generics \code{summary} and
\code{plot}, not used here.}
}
\value{
An \code{fd_data} object.
}
\description{
A fluorescence dilution dataset is a \code{data.frame} of fluorescence
histograms and Colony Forming Units (CFUs)/cell counts for various
timepoints.  It can contain additional attributes that can be used to
simplify the construction of the finite mixture model (FMM) or proliferation
model when calling \code{\link{fd_model}} or \code{\link{fd_unif}}.
\code{fd_data} ensures the conformity of such a dataset by performing various
checks.  Subsetting and \code{rbind}-ing are also available for an
\code{fd_data}, as well as various transformation and display mechanisms.
}
\details{
\code{fd_data} is used to construct and check the validity of a fluorescence
dilution dataset.  \code{fd_transform} applies transformations to histograms
and cell counts and change the attributes \code{"fmm"} accordingly.
Transformations are not chained: the dataset is back transformed using the
current (or default!) transformation, then the new transformation is applied.
\code{cutoff} returns a subset of the dataset where the points below the
cutoff have been discarded.  Notice that \code{subset} cannot be used instead
of \code{cutoff} as the implementation for \code{fd_data} checks that the
histograms sum to one.  \code{fd_expand} uses \code{predict} to fill in more
predicted timepoints between the experimental timepoints (interpolation),
usually for plotting purposes.  \code{plot} can be used for a quick look at
the dataset in order to get an overall "feel" and spot potential mistakes.
}
\section{Format}{
 \code{data} should be a \code{data.frame} (or a
  \code{groupedData}) with at least the following columns:

  \describe{ \item{\code{Type}}{The type of the measurement.  Current allowed
  types are \code{"hists"} (histogram of living cells), \code{"hists_lost"}
  (histogram of dying/dead cells), \code{"Ns"} (cell counts/CFUs),
  \code{"Ns_lost"} (cell counts of dying/dead cells), \code{"props"}
  (proportions of living cells in any given number of generations),
  \code{"props_lost"} (same for dying/dead cells). In the case of histograms,
  one row represents a bin.  However, other types can added by the user,
  provided that a new processing function is given to \code{\link{fd_model}}
  through the argument \code{process}.}

  \item{\code{Weight}}{Either \code{"hist"} (for types \code{"hists"} and
  \code{"hists_lost"}), \code{"N"}, \code{"prop"} or any other weight defined
  by the user (with the right processing function, see argument
  \code{process} of \code{\link{fd_model}}.}

  \item{\code{Inoculum}}{An identifier for the inoculum/progenitors used.
  This allows to associate the histograms with the geometric mean and
  standard deviation of the initial population. Non-histogram based types can
  use \code{"none"} for this field.}

  \item{\code{Timepoint}}{A unique identifier to group together all the
  measurements made together on the same sample.  Notice that if you do
  technical replicates in flow cytometry, the resulting histograms must be
  listed as different timepoints as within the same timepoint there would be
  no way to know which bin pertains to which histogram.}

  \item{\code{Time}}{The time, in hours, at which the timepoint were taken.}

  \item{\code{Category}}{The category/compartment (a factor) for the current
  row.}

  \item{\code{a}, \code{b}}{The left (excluded) and right limits (included)
  of the bin if \code{Type \%in\% c("hists", "hists_lost")}, in linear units
  of fluorescence.  If \code{Type \%in\% c("props", "props_lost") }, \code{a}
  is the number of generations (starting from 0) and \code{b} is ignored. In
  the case of a different type, should NOT be \code{NA} as the whole row can
  be interpreted as missing data by the fitting algorithm or the fitting
  fails if the \code{na.action} is \code{\link[stats]{na.fail}}.  We
  recommend in this case to set \code{a} and \code{b} to 0, as they won't be
  read out by the algorithms anyway.}

  \item{\code{y}}{If \code{Type \%in\% c("hists", "hists_lost") }, this is
  the proportion of cells in bin \code{(a, b]} (right-closed interval, the
  default of \code{\link[graphics]{hist}}). This is a probability and not a
  density and all the rows for a given histogram should sum to one (this is
  checked by this function).  Alternatively, if \code{Type \%in\% c("Ns",
  "Ns_lost") }, this is the cell count/CFU count and if \code{Type \%in\%
  c("prop", "prop_lost") }, this is the proportion of cells in a given number
  of generations. The \code{y} can be transformed, see below the
  \code{cctrans} and \code{htrans} parameters passed to the \code{FMM}. } }

  The following \code{\link[base]{attributes}} can also be set (but setting
  them is not mandatory, default values are provided): \describe{

  \item{\code{fmm}}{List the parameters that will be passed to the finite
  mixture model (FMM), see \code{\link{finite-mixture}}.}

  \item{\code{proliferation}}{List the parameters that will be passed to the
  proliferation model, see \code{\link{proliferation}}.}

  \item{\code{cutoff}}{A named numeric vector containing the cutoff points in
  linear units of fluorescence (the names are the timepoint identifiers).
  See \code{\link{cutoff}}.}

  \item{\code{counts}}{If the user wants to store the number of events per
  histogram, they are advised to use this field.  It should then be a named
  numeric vector (the names would be the timepoint identifiers).}

  }
}

\examples{
# FdSTyphimuriumWTC57 is already an fd_data, so the following
# does not do much
data(FdSTyphimuriumWTC57)
dat <- fd_data(FdSTyphimuriumWTC57)

# Visually assess the data
plot(dat)

# It is possible to perform standard operations on an fd_data,
# as you would do with a data.frame, but with additional checks
datsub <- subset(dat, Individual == "160408.WT.C57")
datbind <- rbind(datsub, subset(dat, Individual == "010708.WT.C57"))

# To disable checks, use e.g. subset.data.frame
invisible(subset(dat, y > 0.1))
  # Warning: some histograms do not have proportions
  # that sum to 1
invisible(subset.data.frame(dat, y > 0.1))
  # Silently returns a data.frame (not an fd_data anymore)

# Cutoff returns an fd_data for which histograms do not have
# proportions that sum to 1, without issuing any warning
invisible(cutoff(dat))

# Sometimes, it is useful to extend the time domain for, e.g.,
# plotting
length(unique(dat$Time))
length(unique(fd_expand(dat, length.N = 50))$Time)
# This expanded dataset can then be used as the 'newdata' argument
# of 'predict'.

# fd_transform can be used for arbitrary transformation
# Notice that here Ns are already log10-transformed
# (this is the default transform as defined in FMM)
dat_tr <- fd_transform(fd_simulate(fd_unif(), c(2, 12, 24)),
                       hist = "log1p")
}
