% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fd_minuslogl}
\alias{fd_minuslogl}
\title{Minus log-likelihood of a fluorescence dilution model.}
\usage{
fd_minuslogl(model, data, start = NULL, mgen = "guess", verbose = TRUE,
  weights = NULL, stop_boundary = FALSE)
}
\arguments{
\item{model}{An \code{\link{fd_model}} object.}

\item{data}{An \code{\link{fd_data}} dataset.}

\item{start}{Optional starting parameters.  The default behaviour
(\code{NULL}) is to use the starting parameter provided by
\code{\link{start}(model)}.}

\item{mgen}{Maximum number of generations to fit.  If \code{NULL}, what the
proliferation model provides by default is used.}

\item{verbose}{If \code{TRUE} (default), print additional diagnostic
messages.}

\item{weights}{A numeric vector of length equal to the number of rows of
\code{data}. Gives the relative weights to put on the residuals.  Default
(\code{NULL}) to equal weight.}

\item{stop_boundary}{Whether \code{\link{fd_predict}} should raise an error
if the parameters cross the boundary.  By default (\code{FALSE}), just
returns a very large number when it happens.}
}
\value{
Minus log-likelihood (same as \code{\link[stats]{logLik}} applied to
  an \code{\link{nls}} object).
}
\description{
This function returns a minus log-likelihood function to be used with general
maximum likelihood estimation packages such as \pkg{stats4} or \pkg{bbmle}.
}
\examples{
library(stats4)
library(bbmle)

# Create minuslogl function
data(FdSTyphimuriumWTC57)
dat <- subset(cutoff(FdSTyphimuriumWTC57), Individual == "140508.WT.C57")
mdl <- fd_model(dat, fmm="gaussian", proliferation="branching",
                constraints = attr(dat, "bestcstr"))
fun <- fd_minuslogl(mdl, dat, verbose=FALSE)
print(fun)

# This this function on lower parameters
do.call(fun, as.list(start(mdl)))

# Optimize this minuslogl.  Notice that the 'start' parameter is not
# explicitly given as the arguments to 'fun' have default values
fun <- fd_minuslogl(mdl, dat, verbose=TRUE)
control <- list(maxit=1)
\dontrun{
control <- NULL
}
stats4::mle(fun, method="Nelder-Mead", control=control)

# The user is advised to use GenSA for global search:
bbmle::mle2(fun, optimfun=GenSA::GenSA, control=control)
}
\seealso{
Other optimization-related entries: \code{\link{fd_comb}},
  \code{\link{fd_gaussian_fmm_solve}}, \code{\link{fd_nls}}
}
