% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{fd_peaks}
\alias{fd_peaks}
\alias{fd_findpeaks}
\alias{fd_peaks}
\title{Locate the distinct peaks in fluorescence dilution histograms.}
\usage{
fd_findpeaks(data, plot = TRUE, w = 5, span = 0.1)

fd_peaks(data, peaks, rm = TRUE)
}
\arguments{
\item{data}{An \code{\link{fd_data}} dataset.}

\item{plot}{Whether a diagnostic plot of the peaks should be displayed.}

\item{w}{Window parameter: how large (in number of histogram bins) should the
window used to calculate the maximum be?}

\item{span}{The parameter that controls the degree of smoothing in
\code{\link[stats]{loess}}.}

\item{peaks}{The return value of an \code{fd_findpeaks} call, with an
additional \code{Generation} column mapping the peaks to numbers of
generations.}

\item{rm}{Whether the rows with \code{Weight=="hist"} should be removed in
the returned \code{\link{fd_data}} object.}
}
\value{
\code{fd_findpeaks} and \code{fd_peaks} both return a
  \code{data.frame} (see \emph{details}).
}
\description{
Find the peaks, using a smoothing method, of fluorescence
histograms.  This could be useful in order to eschew the \emph{a priori}
assumptions concerning the evolution over time and generations of the space
between those peaks that finite mixture models make (Hyrien et al.
(2008) have showed that the "naive", Gaussian FMM is sometimes
inappropriate).  However, clear peaks are not discernible in microbiology
applications so far, restricting this technique to immunological
applications.
}
\details{
\code{fd_findpeaks} returns a \code{data.frame} giving the
\code{"Timepoint"}, position of the peak \code{"x"}, proportion of cells in
the peak \code{"prop"} and position of the peaks from right to left
\code{"i"}, starting from 1. Notice that the number of generation is not
directly given as sometimes very low proportions could mean that a generation
is skipped and not recorded as a peak.  Since this should not pose any
problem, apart from labelling, the labelling should be done by the user (see
the examples section for a straightforward simple mapping).

\code{fd_peaks} takes such a mapping and appends to the dataset the
proportions, with weight \code{"prop"} and types \code{"props"}
(corresponding to \code{"hists"} histograms) and \code{"props_lost"}
(corresponding to \code{"hists_lost"} histograms).
}
\section{Required packages}{
 Please install \pkg{zoo}.
}

\examples{
data(FdClearPeaks)
dat <- FdClearPeaks

# The above dataset has a low noise level, therefore
# no transformation is really needed to map peaks to
# generation numbers
peaks <- transform(fd_findpeaks(dat, span=0.05, w=6),
                   Generation = i - 1)
fd_peaks(dat, peaks, rm=TRUE)
}
\seealso{
\code{\link{finite-mixture}}, \code{\link{fd_gaussian_fmm_solve}}.
}
