% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_model.R
\name{fd_predict}
\alias{fd_predict}
\title{Predict the histograms and cell counts of an \code{fd_data} dataset from a
set of model parameters.}
\usage{
fd_predict(model, param, data, mgen = NULL, stop_boundary = FALSE)
}
\arguments{
\item{model}{An \code{\link{fd_model}} object.}

\item{param}{A named vector of free parameters, such as returned by
\code{\link{start}}.}

\item{data}{An \code{\link{fd_data}} dataset.}

\item{mgen}{Maximum number of generations.  If \code{NULL}, taken from the
default found in the proliferation model.}

\item{stop_boundary}{Whether the function should stop when \code{param} is
out of bounds or simply return very large predicted values.}
}
\value{
The original \code{fd_data}, not necessarily in the same order, with
  the \code{"y"} column filled with the predicted values.
}
\description{
Predict the histograms and cell counts of an \code{fd_data} dataset from a
set of model parameters. \code{fd_predict} is used internally when
\code{\link{fd_formula}}, \code{\link{fd_simulate}} or
\code{\link{fd_minuslogl}} are called, but it can also be called directly.
}
\examples{
library(dplyr)

data(FdSTyphimuriumWTC57)
dat <- cutoff(FdSTyphimuriumWTC57)
dat <- dat[order(dat$Type, dat$Inoculum), ]
mdl <- fd_model(data=dat, constraints=attr(dat, "bestcstr"))
pred <- fd_predict(mdl, start(mdl), dat)
if (any(all.equal(pred \%>\% dplyr::select(-y),
                  as.data.frame(dat) \%>\% dplyr::select(-y)) != TRUE))
  stop("should be equal")
dat_residuals <- pred$y - dat$y
}
\seealso{
Other simulation-related entries: \code{\link{dataset-simulation}},
  \code{\link{random-parameters}}
}
