% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluodilution.R
\docType{data}
\name{FdClearPeaks}
\alias{FdClearPeaks}
\title{Artificial example dataset for one-compartment fluorescence dilution.}
\format{This is an \code{\link{fd_data}} with the FMM and proliferation
  parameters preset using attributes.}
\usage{
data(FdClearPeaks)
}
\description{
This dataset was generated by \code{\link{fd_simulate}} and features clear,
distinct peaks such as those found in immunological applications.
}
\details{
This dataset can be best fitted with a Gaussian finite mixture model
(\code{\link{fd_fmm_gaussian}}) and a branching model
(\code{\link{fd_proliferation_branching}}).
}
\examples{
data(FdClearPeaks)
plot(FdClearPeaks)
}
\seealso{
\code{\link{fd_data}} for data format, \code{\link{fd_model}} for an
  example of use.
}
\keyword{datasets}
