% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluodilution.R
\docType{data}
\name{FdSTyphimuriumWTC57}
\alias{FdSTyphimuriumWTC57}
\title{Example dataset for one-compartment fluorescence dilution.}
\format{This is an \code{\link{fd_data}} with the FMM and proliferation
  parameters preset using attributes.}
\source{
Sophie Helaine, Jessica A. Thompson, Kathryn G. Watson, Mei Liu,
  Cliona Boyle, and David W. Holden (2010).  Dynamics of intracellular
  bacterial replication at the single cell level.  \emph{Proc Natl Acad Sci
  USA.} \strong{107} (8):3746-3751.
}
\usage{
data(FdSTyphimuriumWTC57)
}
\description{
In this experiment, primary BMM extracted from C57 B1/6 mice were infected by
wild-type 12023s \emph{Salmonella} Typhimurium harbouring the pFCcGi plasmid,
a plasmid with a constitutive reporter helping in bacterial recognition and
an inducible reporter used for fluorescence dilution.  The data was acquired
on a LSRFortessa cytometer (DB).  Raw data is available upon request.
}
\details{
This dataset can be best fitted with a Gaussian finite mixture model
(\code{\link{fd_fmm_gaussian}}) and a branching model
(\code{\link{fd_proliferation_branching}}) with the common constraints\verb{
}\code{~ `#noss` + `#nodeath` + `#delta_1100` + `#mm_xyxz`} (see
\code{\link{FdCommonConstraints}}).  As this was a repeat experiment (the
field \code{Individual} gives an identifier for the experiment), it is also
possible to explore the use of mixed-effects models \emph{via}
\code{\link[nlme]{nlme}}.
}
\examples{
data(FdSTyphimuriumWTC57)
plot(FdSTyphimuriumWTC57)
}
\seealso{
\code{\link{fd_data}} for data format, \code{\link{fd_model}} for an
  example of use.
}
\keyword{datasets}
