% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma-distributions.R
\name{gamma-distributions}
\alias{gamma-distributions}
\alias{fd_gamma_dist}
\alias{fd_reparam_gamma_dist}
\alias{fd_pack_dist}
\alias{fd_pdist}
\alias{fd_ddist}
\title{Reparametrization and evaluation of shifted gamma distributions.}
\usage{
fd_gamma_dist(a, b, loc = 0, mul = 1)

fd_reparam_gamma_dist(mm, delta = 1, ss = 0.5)

fd_pack_dist(dst, mul = 1)

fd_pdist(t, g)

fd_ddist(t, g, dt = NULL)
}
\arguments{
\item{a}{The shape \eqn{\alpha} of the gamma distribution.}

\item{b}{The rate \eqn{\beta} of the gamma distribution.}

\item{loc}{The location parameter \eqn{\lambda}.}

\item{mul}{A multiplier factor.  Used internally for representing mixtures.}

\item{mm}{The mean of the gamma distribution, in hours.}

\item{delta}{The coefficient of variation (i.e., standard deviation over
mean) of the gamma distribution.  It is dimensionless and should lie
between 0 and 1.}

\item{ss}{The reparametrization of the location parameter.  It is
dimensionless and should lie between 0, excluded (high skewness as compared
to a non-shifted gamma distribution), and 0.5, included (same skewness as
compared to a shifted gamma distribution).}

\item{dst}{A list of the parameters of a gamma distribution in their
reparametrized form: \code{dst$mm}, \code{dst$delta}, \code{dst$ss}.}

\item{t}{A numeric vector of times in hours at which to evaluate the gamma
distribution.}

\item{g}{A gamma distribution in the natural form (output of
\code{fd_gamma_dist}).}

\item{dt}{A list of small time increments for evaluating the gamma
distribution (see details).}
}
\value{
See \emph{details}.
}
\description{
Our proliferation models use gamma distributions for times to division/death.
They are parametrized by their mean, coefficient of variation and a
parametrization of the skewness. These functions help in the manipulation of
such distributions.
}
\details{
Gamma distributions are usually parametrized using scale and shape
parameters. Additionally, a shifted gamma distribution makes use of a
location parameter. It is possible to use a more "natural" parametrization
from the mean, coefficient of variation and an alternative reparametrization
of the location parameter (we could think of it as a \emph{second-order}
shape).

\code{fd_gamma_dist} takes the parameters of a shifted gamma
distribution in their non-reparametrized (natural) form and returns a one-row
matrix with those parameters. \code{fd_reparam_gamma_dist} just returns a
list made with its parameters, in the reparametrized form.
\code{fd_pack_dist} takes a repametrized form and returns a natural form in
the format of \code{fd_gamma_dist}.

\code{fd_pdist} evaluates the cumulative distribution function of the gamma
distribution \code{g} (natural form) at times \code{t}.  \code{fd_ddist}
evaluates the density function \eqn{g(t)} when \code{dt == NULL}.  When
\code{dt} is a numeric vector, it evaluates the small increments \eqn{g(t) \
dt}{g(t) dt} (so \code{dt} must have the size of \code{t}).
}
\section{(Shifted) gamma distributions}{
 Times to division/death can be
  parametrized by any continuous distribution.  However, as observed by
  Hawkins et al. (2007) and Miao et al. (2012), a gamma
  distribution is often flexible enough.  A gamma distribution, contrary to a
  normal distribution, has a nonnegative support and is skewed, two
  properties that are expected of distributions of times to division/death
  (Hyrien et al. (2008)).  To ascertain whether finer parametrization
  can lead to better results, we allow the user to set independently not only
  the mean and variance of the distribution, but also its skewness (third
  moment) through a distribution \emph{shift}.

  More specifically, is possible to extend the gamma distribution by shifting
  it by a location parameter \eqn{\lambda \ge 0} (type III distribution in
  the Pearson Distribution System): \eqn{\Gamma_{shifted}(t; \delta, \bar
  \tau, \lambda) = \Gamma(t - \lambda; \delta, \bar \tau) }{\Gamma/shifted(t;
  \delta, mean \tau, \lambda) = \Gamma(t - \lambda; \delta, mean \tau)} (with
  the convention that \eqn{\Gamma(t) = 0} for \eqn{t < 0}). (The gamma
  distribution is scale-invariant, so it would not make sense to add a
  fourth, scale parameter as is usually done to get location-scale
  families.)   To get relevant bounds and help convergence, we can
  reparametrize the distribution as follows.  Noting \eqn{\gamma > 0} the
  skewness, \eqn{\Delta > 0} the coefficient of variation and \eqn{\mu > 0}
  the mean, we have \deqn{\alpha = \frac{4}{\gamma^2} \quad\quad \beta =
  \frac{2}{\mu \gamma \Delta} \quad\quad \lambda = \mu (1 - 2 \Delta /
  \gamma)}{ \alpha = \gamma^2 / 4; \beta = \mu \gamma \Delta / 2.0; \lambda =
  \mu (1 - 2 \Delta / \gamma)}

  We clearly see with this last equality that we must have \eqn{\Delta /
  \gamma \le 1/2}: instead of parametrizing with \eqn{\gamma}, we parametrize
  with the ratio \eqn{0 < u = \Delta / \gamma \le 1/2} of the coefficient of
  variation over the skewness.
}

\section{Special values}{
 All the proliferation models in this package allow
  the special, "degenerate" value \code{delta = 0}.  In this case, the
  distribution of times to division/death is a Dirac function with the mean
  still given by \code{mm} and \code{ss} is irrelevant. Moreover, because the
  distribution does not have a density anymore, \code{fd_ddist} must be
  called with \code{dt} non NULL (it gives a numeric vector filled with
  \code{0} except for one element). Also, \code{fd_pdist} reduces to a step
  function.
}

\examples{
mdl <- fd_model()
dst <- fd_pack_dist(relist(start(mdl), mdl)$pro$One$f)
curve(fd_pdist(x, dst), from=0, to=20)
curve(fd_ddist(x, dst), from=0, to=20, add=TRUE, col="red")

}
\references{
Hawkins ED, Turner ML, Dowling MR, van Gend C, Hodgkin PD (2007).
A model of immune regulation as a consequence of randomized lymphocyte
division and death times. \emph{Proc Natl Acad Sci USA} \strong{104} (12):
5032-5037.

Hyrien O, Zand MS (2008). A Mixture Model with Dependent Observations for the
Analysis of CSFE-Labeling Experiments. \emph{Journal of the American
Statistical Association} \strong{103} (481): 222-239.

Miao H, Jin X, Perelson AS, Wu H (2012). Evaluation of multitype mathematical
models for CFSE-labeling experiment data. \emph{Bull Math Biol} \strong{74}
(2): 300-326.
}
