% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsel.R
\name{model-selection}
\alias{model-selection}
\alias{fd_aictab}
\alias{fd_modavg}
\alias{vcov.fd_modavg}
\alias{weights.fd_modavg}
\alias{best}
\title{Model selection based on an \code{\link{fd_comb}} object.}
\usage{
fd_aictab(second.ord = FALSE, data, ..., .list = NULL)

fd_modavg(tab, vcov. = vcov)

\method{vcov}{fd_modavg}(object, ...)

\method{weights}{fd_modavg}(object, ...)

best(tab, subset = TRUE, pos = 1, verbose = TRUE)
}
\arguments{
\item{second.ord}{Whether to use AIC (\code{FALSE}) or AICc (\code{TRUE}).
We discourage the use of AICc as this has been insufficiently dealt with
within the statistical literature on generalized estimating equations.  If
\code{TRUE}, the number of observations used for the small-sample
adjustment is the number of clusters.  This clearly makes the adjustment
bigger than it should be.}

\item{data}{The \code{\link{fd_data}} dataset on which optimization was
performed.}

\item{...}{(Potentially named) \code{\link{fd_nls}}/\code{link[stats]{nls}}
objects.}

\item{.list}{Alternatively or additionally, a (potentially named) list of
\code{\link{fd_nls}}/\code{link[stats]{nls}} objects can be provided.}

\item{tab}{An \code{fd_aictab} object.}

\item{vcov.}{Variance-covariance function to use on the free parameters. For
example, the \pkg{sandwich} package offers an alternative to
\code{\link[stats]{vcov}} and could be used for a Huber-White estimation.
Alternatively, variance-covariance matrix.}

\item{object}{An \code{fd_modavg} object.}

\item{subset}{Used to subset \code{tab} the same way the \code{subset}
argument of \code{\link[base]{subset}} works.}

\item{pos}{If \code{1}, returns the best model.  If \code{2}, returns the
second-best model, etc.}

\item{verbose}{whether the name of the chosen model should be explicitly
printed.}
}
\value{
\code{fd_aictab} returns a \code{data.frame} (see \emph{details}),
\code{fd_modavg} an \code{fd_modavg} object, \code{best} a "fit" object
(e.g., an \code{\link{fd_nls}} object).
}
\description{
Wrappers around the functions of package \pkg{AICcmodavg} and extensions
inspired by it specifically tailored to nonlinear optimization.
}
\details{
\code{fd_aictab} returns a \code{data.frame} of various summary
statistics concerning the models, ranked by increasing AIC or AICc.  The
output builds on the output of \code{\link[AICcmodavg]{aictab}}, with two
additional columns: \code{Effect} and \code{Q} (see the section on "effect
size" below for interpretation).

The \code{fd_modavg} function works similarly to
\code{\link[AICcmodavg]{modavg}} function. It is based on a conceptualization
of AIC values as weights (Burnham and Anderson 2004). The weighting is done
both on the estimate and on the variance-covariance matrix.  \code{coef},
\code{vcov} and \code{weight} can be used to extract, respectively, the
estimated model-averaged coefficients, the variance-covariance matrix and the
weights used for the averaging.  Using QIC for correcting for variance
structure misspecification (Pan 2001) would have been better, but
implementation in R is sketchy and this is a fundamental matter beyond the
scope of this work.

\code{best} returns the best (or second-best, ...) model of an
\code{fd_aictab} object.
}
\section{Effect size}{
 The additional columns \code{Effect} and \code{Q} are
  to be interpreted as follows.

  We think of a parameter as practically identifiable when its 95\%
  confidence interval does not include its lower or upper bounds.  This
  translates into a normalized effect size of roughly \eqn{2} (\eqn{p}-value
  < \eqn{0.05}, \code{*}).  Effect sizes greater than \eqn{2.6} (\eqn{p <
  0.01}, \code{**}) or even \eqn{3.3} (\eqn{p < 0.001}, \code{***}) are even
  more desirable. We think of a model as practically identifiable when the
  effect size of each of its parameters that have not reached the lower
  bounds is at least 2, and in general we quantify the practical
  identifiability of a model with the minimum of the effect sizes (we simply
  call it the \code{Effect} of the model).  A qualitative appraisal of this
  effect is given by \code{Q}.  Its value is either \code{""} (\code{Effect <
  2}), \code{"*"}, \code{"**"} or \code{"***"} depending on the minimum
  effect size.
}

\section{Required packages}{
 Please install \pkg{AICcmodavg} for
  \code{fd_aictab}.
}

\examples{
control <- list(maxit=1)

\dontrun{
control <- NULL
}

data(FdSTyphimuriumWTC57)
dat <- cutoff(FdSTyphimuriumWTC57)
source(system.file("contrib", "nlsSA.R", package="fluodilution"))
CC <<- FdCommonConstraints
comb <- fd_comb(catcstr(attr(dat, "bestcstr"),
                        list(CC$`#delta_1100`, CC$`#delta_1111`)),
                data=dat,
                nlsfun="nlsSA",
                control=control)

tab <- fd_aictab(data = dat, .list = comb)
avg <- fd_modavg(tab)
invisible(vcov(avg))
weights(avg)
best(tab)

}
\references{
Burnham KP, Anderson DR (2004). Multimodel inference:
understanding AIC and BIC in model selection. \emph{Sociological Methods
Research} \strong{33}: 261-304.

Pan W (2001). Akaike's Information Criterion in Generalized Estimating
Equations. \emph{Biometrics} \strong{57} (1): 120-125.
}
