% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation-.R, R/proliferation-branching.R,
%   R/proliferation-cyton.R
\name{proliferation}
\alias{proliferation}
\alias{fd_proliferation_branching}
\alias{fd_proliferation_cyton}
\title{Implementation of the Cyton and branching models.}
\usage{
fd_proliferation_branching(...)

fd_proliferation_cyton(...)
}
\arguments{
\item{...}{\describe{ \item{\code{categories}}{Name of categories for the
proliferation model (branching processes only).} \item{\code{mgen}}{Maximum
number of generations to model.} \item{\code{length.out}}{Number of time
points at which division/death rates are evaluated.  The more the better,
but also the slower (Cyton models only).} \item{\code{log10_scale}}{Whether
to reparametrize, when fitting, proportions (p, p0, res/res0,
mrates/mrates0) using a \code{log10} scale in order to potentially speed up
convergence (branching processes only).} \item{\code{initial}}{A matrix
giving the initial relative number of seeds in each category (rows) and for
each number of generations (columns).  Categories are given in the order of
\code{levels(data$Category)}, with \code{data} an \code{\link{fd_data}}
object and generations from 0 to \code{mgen} (included).  If a different
\code{mgen} is chosen for \code{\link{fd_formula}}, the matrix is
automatically resized and padded by zeroes if necessary.} }}
}
\value{
An \code{fd_proliferation} object.
}
\description{
With this package, a large variety of proliferation models can
be implemented.  They all rest on two broad sets of assumptions. On one hand,
a generalized Cyton model and its nested, simpler models, consider that
division and death occur concurrently, in competition (Hawkins et al.
2007). On the other hand, a branching process (Hyrien et al. 2010)
considers that birth and death have both a given probability of occurring in
a cell (probability of transition) and a given probability of occurring after
a certain time (time to transition).
}
\section{Parametrization}{
 A proliferation model is parametrized as a named
  list with as many elements as there are compartments/categories.  The order
  in which those elements are arranged has its importance: because no
  feedback is allowed, migration can only happen from a compartment with a
  lesser index to a compartment with a greater index.  Moreover, although
  with Cyton models migration is not implemented, for symmetry the parameters
  are bundled in a list with one element named \code{"One"}.

  The shared parameters within those elements, for both models, are as
  follows (be careful, they however DO NOT share the exact meaning, see
  Chauvin et al. (2016)): \describe{ \item{\code{f0$mm, f0$delta,
  f0$ss}}{Parametrization of the times to first division. See
  \code{\link{fd_gamma_dist}} for details.} \item{\code{f$mm, f$delta,
  f$ss}}{Parametrization of the times to subsequent divisions.}
  \item{\code{g0$mm, g0$delta, g0$ss}}{Parametrization of the times to death
  before the first division.} \item{\code{g$mm, g$delta,
  g$ss}}{Parametrization of the times to death after the first division.}
  \item{\code{res0, res0}}{Proportion of cells that never divide nor die
  (\code{res0}) and that do not divide further nor die after the first
  generation (\code{res}).} }

  Additionally, \itemize{ \item{In the natural parametrization (see
  \code{\link{fd_model}}), \code{fd_proliferation_branching} accepts
  \code{p0} and \code{p}, the proportion of cells that undergo first division
  and undergo division in any number of generations greater than 0, and
  \code{mrates}, which is a named list of migration "rates" between 0 and 1L
  for instance, \code{Two$mrates$One} gives the migration "rate" from
  compartment "One" to "Two".} \item{For branching processes only, additional
  parameters \code{h$mm, h$delta, h$ss} can specify a decay distribution,
  giving times to disintegration for dying/dead cells, and
  \code{"hist_lost"}, \code{"props_lost"} and \code{"Ns_lost"} types (see
  \code{\link{fd_data}}) take this into account.  The default constraining
  does not specify \code{h} and as a consequence no decay is used. This
  decaying is as of now not available for Cyton models.  Notice moreover
  that, as of now, decaying is not implemented properly for
  multicompartmental models as decay would also impact emigrating cells.} }
}

\section{Reparametrization}{
 For \code{fd_proliferation_branching}, if
  \code{log10_trans} is set to \code{TRUE}, proportions are reparametrized
  (see \code{\link{fd_model}}) by their logarithm in base 10. Moreover,
  \code{p0} and \code{p} are the proportions of cells that do not die at the
  current generation, either because they divide or are nongrowers, and
  \code{res0} and \code{res} are given relative to \code{p0} and \code{p}:
  therefore, for branching processes the proportion of nongrowers, among the
  total number of cells in a given number of generations, is \code{p0 * res0}
  or \code{p * res}.

  For \code{fd_proliferation_cyton}, no reparametrization is undertaken.
}

\section{Special cases}{
 Notice that proliferation models allow the
  \code{delta} to be 0, in which case the respective distributions reduce to
  Dirac distributions, and the \code{g$mm} and \code{g0$mm} to be infinite
  (\code{Inf}), in which case no death occurs (see
  \code{\link{gamma-distributions}}).
}

\examples{
fd_model(proliferation="cyton")
fd_model(proliferation=fd_proliferation_branching(c("One", "Two"), mgen=10))
fd_proliferation_branching(c("One", "Two"), mgen = 10L,
                           initial = matrix(c(1, 0.5, 0, 0), nrow=2))
}
