% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_simulate.R
\name{random-parameters}
\alias{random-parameters}
\alias{fd_draw_unif}
\alias{fd_unif}
\alias{fd_norm}
\title{Generate random parameters.}
\usage{
fd_draw_unif(model, n = 1, coverage = 0.8)

fd_unif(fmm = "gaussian", proliferation = "branching", constraints = NULL,
  n = 1, coverage = 0.8)

fd_norm(n, mean, cv = 0.1, model = NULL)
}
\arguments{
\item{model}{An \code{\link{fd_model}} object.  In the case of
\code{fd_norm}, mandatory only if \code{mean} was not generated by
\code{fd_unif}/\code{fd_draw_unif}.}

\item{n}{Number of sets of parameters to draw.}

\item{coverage}{A number between \code{0} and \code{1} indicating how much of
the lower bounds-upper bounds interval should be covered by the random
draw: for example, a coverage of \code{0.8} means that the interval will be
shrank symmetrically to 80\% of its size.  This avoids drawing too extreme
values.}

\item{fmm}{A Finite Mixture Model (FMM), see \code{\link{finite-mixture}}.}

\item{proliferation}{A proliferation model, see \code{\link{proliferation}}.}

\item{constraints}{A \code{cstr} object, as produced by
\code{\link{cstrlist}} or an \code{expression} if only the
\code{constraint} component is set (see \code{\link{constraints}}).}

\item{mean}{A named numeric vector of coefficients or a matrix with parameter
names in columns as would be returned by \code{coef} applied to an
\code{\link[nlme]{nlme}}, \code{\link[nlme]{nlsList}},
\code{\link[stats]{nls}} or \code{\link{fd_nls}} object or the result of
\code{\link{fd_unif}}.  If it is a matrix, the columns are simply averaged.}

\item{cv}{Coefficient of variation (equal across all coefficients) of the
normal distribution from which to draw.}
}
\value{
A named numeric vector of free parameters (\code{n=1}) or a matrix
  (\code{n>1}), augmented by information about the model that can be queried
  with \code{\link{model}}.  Therefore, it is possible to use
  \code{\link{relisted_coef}} on the return value.
}
\description{
Generate random parameters from an \code{\link{fd_model}} with
  either a uniform draw (\code{fd_draw_unif} or the shortcut \code{fd_unif})
  or distributed normally (\code{fd_norm}).
}
\section{Required packages}{
 Please install \pkg{truncnorm} for \code{fd_norm}
}

\examples{
(ans <- fd_unif("gaussian", "branching"))
fd_draw_unif(model(ans))
fd_norm(3, ans)

}
\seealso{
Other simulation-related entries: \code{\link{dataset-simulation}},
  \code{\link{fd_predict}}
}
